only on b6-rk3288-firefly-4gb

+++ tools/bpo/build_deb.sh
@@ -5,6 +5,7 @@
 sudo mount -a
 
 localdir="/mnt/farm"
+builder=`cat /etc/hostname`
 
 build () {
 	if [ "x${sbuild_chroot}" = "x" ] ; then
@@ -43,7 +44,9 @@ cleanup_suite () {
 }
 
 run () {
-	touch /tmp/chroot-BUILDING.lock
+	if [ "x${builder}" = "xb6-rk3288-firefly-4gb" ] ; then
+		touch /tmp/chroot-BUILDING.lock
+	fi
 	out_dir="${localdir}/outgoing/${suite}/${deb_arch}/${debian_pkg_name}_${debian_version}"
 	if [ -f /var/lib/sbuild/${suite}-${deb_arch}.tar.gz ] ; then
 
@@ -56,7 +59,9 @@ run () {
 
 		cd ../
 	fi
-	rm -f /tmp/chroot-BUILDING.lock || true
+	if [ "x${builder}" = "xb6-rk3288-firefly-4gb" ] ; then
+		rm -f /tmp/chroot-BUILDING.lock || true
+	fi
 }
 
 if [ ! "x${dist}" = "x" ] ; then
